test: fix bug in test and update test manifest
This fixes a bug in the test where if multiple files are to
be validated by the test, the test incorrectly `returns` from
a `for` loop that asserts the correctness of multiple expected files
when the first expected file succeeds validation. As a result, the
manifest for testpkg2.json was never updated as the code evolved,
and is incorrect in its current form.

Signed-off-by: Shashank Ram <shashank.ram@solo.io>